Document Very Minimal changes
To show the changes made from default EdyTheCow's docker-pterodactyl files to get it working for me.

Readme can be somewhat ignored in most parts, was created to help a lesser experienced member of my team setup a node himself.

Doesn't include information about the changes to the panel. The same steps to setup the traefik for the node are also used for the panel to avoid having to supply two sets of the traefik composes.